Rename dynamicPages folder to templates as this matches the Gatsby file structure
@see https://www.gatsbyjs.org/docs/gatsby-project-structure/
