+++ src/templates/platformDetailDataProvider.tsx
